+++ docs/changelog.md
@@ -10,6 +10,37 @@ pw-framework 的版本号遵循 [语义化版本 2.0.0](https://semver.org/lang/
 先行版本号及版本编译元数据可以加到“主版本号.次版本号.修订号”的后面，作为延伸。
 ```
 
+# v21.2.1（2019/3/20）
+
+- 修复 `TableData` 组件中后端修改按钮记录表单不显示字段值的 bug
+- 修复 `TableData` 组件中记录表单的 title 显示错误的 bug
+
+# v21.2.0（2019/3/18）
+
+- `TableData` 组件添加 `recordFormFormWidth` 和 `recordFormTabsWidth` 配置
+- 修复在 `TableData` 组件中记录表单在只有右侧 tabs 时，tabs 宽度为 50% 的 bug
+- 修复 `TableData` 在添加记录时，表单宽度为 50% 的 bug
+
+# v21.1.1（2019/3/16）
+
+- 修复在 `TableData` 组件中只有后端按钮时，点击不会弹出记录表单的 bug
+
+```
+*升级提示
+git pull origin master
+```
+
+# v21.1.0（2019/3/15）
+
+- `TableData` 组件添加 `hasSearch` 配置
+- 修复在 `TableData` 组件中点击后端按钮，不会获取记录表单中的子表数据的 bug
+- 修复在 `TableData` 组件中点击后端按钮，请求传参错误的 bug
+
+```
+*升级提示
+git pull origin master
+```
+
 # v21.0.0（2019/3/14）
 
 - `app.config.js` 中添加 `开发环境` 和 `生产环境` 配置